fetchData.js moved into MovieList catalog
